fix default map issue and enable more free-flowing navigation
I believe this fixes #60 the issue where the first custom map drawn wasn't loading, but the second and others would. Now, anytime a custom map is drawn, it will load.

This also addresses #18. I believe at this point, as long as the user lands on welcome, they can navigate in any order they want and they will at least be able to get the default area (but if they don't go to Violations first they won't get markers on Lead and EJ but that could be addressed later on and it isn't vital).
